+++ CHANGELOG.md
@@ -7,9 +7,11 @@
 * (bluefox) Added support for dynamic notification layout (in Admin)
 * (@foxriver76) updated plugin base and sentry plugin to version 2
 * (@foxriver76) enhanced translations for the `diskSpaceIssues` notification category
+* (@foxriver76) added notification if new image is available on Docker Hub (for official docker systems)
 * (@foxriver76) extend the time to wait until controller is stopped on controller UI upgrade
 * (@foxriver76) improved backup/restore process to work for arbitrary large installations
 * (@GermanBluefox/@foxriver76) implemented automatic upload on adapter start if version mismatch is detected
+* (@foxriver76) enriched logging on upload with version information
 
 ## 6.0.11 (2024-08-21) - Kiera
 * (foxriver76) only generate `packageUpdates` notification, if new updates detected